+++ server/src/main/java/org/allaymc/server/block/component/BlockLiquidBaseComponentImpl.java
@@ -96,11 +96,6 @@ public abstract class BlockLiquidBaseComponentImpl extends BlockBaseComponentImp
         return true;
     }
 
-    @Override
-    public boolean canCollideWithEntity() {
-        return true;
-    }
-
     /**
      * This method is used in {@link org.allaymc.server.world.service.AllayEntityPhysicsService}
      */
@@ -247,7 +242,6 @@ public abstract class BlockLiquidBaseComponentImpl extends BlockBaseComponentImp
      *
      * @param dimension The dimension the block is in.
      * @param pos       The position to check around.
-     *
      * @return Whether there is a supply liquid around the position.
      */
     protected boolean hasSupplyLiquidAround(Dimension dimension, Vector3ic pos, BlockState liquid) {
@@ -294,7 +288,6 @@ public abstract class BlockLiquidBaseComponentImpl extends BlockBaseComponentImp
      * @param liquid    The block state of the liquid.
      * @param pos       The position to flow into.
      * @param falling   Whether the liquid is falling or not.
-     *
      * @return Whether the liquid successfully flowed into the position.
      */
     protected boolean flowInto(Dimension dimension, Vector3ic src, int srcLayer, BlockState liquid, Vector3ic pos, boolean falling) {
@@ -377,7 +370,6 @@ public abstract class BlockLiquidBaseComponentImpl extends BlockBaseComponentImp
      * @param src             The position to start the flow from.
      * @param liquid          The block state of the liquid.
      * @param liquidContainer The block state that contains this liquid block, can be {@code null} if the liquid is not contained.
-     *
      * @return A list of paths that the liquid can flow in.
      */
     protected List<Vector3ic[]> calculateLiquidPaths(Dimension dimension, Vector3ic src, BlockState liquid, BlockState liquidContainer) {
@@ -433,7 +425,6 @@ public abstract class BlockLiquidBaseComponentImpl extends BlockBaseComponentImp
      * @param source    The source position of the liquid.
      * @param node      The node to spread.
      * @param queue     The queue to push the node into if it can spread.
-     *
      * @return Whether the node could spread into the neighbour.
      */
     protected boolean spreadNeighbor(Dimension dimension, Vector3ic source, LiquidNode node, LiquidQueue queue) {
@@ -477,7 +468,6 @@ public abstract class BlockLiquidBaseComponentImpl extends BlockBaseComponentImp
      * @param y         The y coordinate of the block.
      * @param z         The z coordinate of the block.
      * @param sideways  Whether the flow is sideways or downwards.
-     *
      * @return Whether the liquid can flow into the block.
      */
     protected boolean canFlowInto(Dimension dimension, int x, int y, int z, boolean sideways) {
@@ -511,7 +501,6 @@ public abstract class BlockLiquidBaseComponentImpl extends BlockBaseComponentImp
          * neighbours returns the four horizontal neighbours of the node with decreased depth.
          *
          * @param decay The amount to decrease the depth by.
-         *
          * @return The neighbours of the node.
          */
         public LiquidNode[] neighbors(int decay) {